# [PERF] use numexpr in dispatch_to_series