@jbrockmendel 2018-08-13
> the perf numbers look about the same.

Yah, this is actually a win though.  A previous attempt to use `dispatch_to_series` instead of `_compare_frame` had a 15x slowdown.  This gets us to a .8x speedup.  But the real motivation is to avoid the redundant dispatch code in `_compare_frame`

BTW for triage purposes, getting #22163 through will make it easier to do a bunch of test cleanup/parametrization.  (and, you know, fix a bunch of bugs)