@jbrockmendel 2018-08-22
The perf part of this PR is really secondary.  The main goal is to unify DataFrame ops going through `dispatch_to_series`.  The numexpr part is necessary to avoid a perf hit in frame/frame comparisons.