@pep8speaks 2018-09-05
Hello @jbrockmendel! Thanks for updating the PR.

 - There are no PEP8 issues in the file [`pandas/core/frame.py`](https://github.com/jbrockmendel/pandas/blob/13016e72ad01df10d65b85a7a5befb050e5702b4/pandas/core/frame.py) !

 - There are no PEP8 issues in the file [`pandas/core/ops.py`](https://github.com/jbrockmendel/pandas/blob/13016e72ad01df10d65b85a7a5befb050e5702b4/pandas/core/ops.py) !